requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
plotly>=5.18.0